xfinity-usage/requirements.txt
typing-extensions==4.12.2
tenacity==9.0.0
requests==2.32.3
PyJWT==2.10.1
paho-mqtt==2.1.0
colorlog==6.9.0
cryptography==44.0.0